+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [3.6.2](https://github.com/Automattic/newspack-ads/compare/v3.6.1...v3.6.2) (2025-08-13)
+
+
+### Bug Fixes
+
+* **placements:** side rail placement dimension race condition ([#984](https://github.com/Automattic/newspack-ads/issues/984)) ([304a9f3](https://github.com/Automattic/newspack-ads/commit/304a9f3163943379bc3887ae4f22fda241151cc7))
+
+## [3.6.1](https://github.com/Automattic/newspack-ads/compare/v3.6.0...v3.6.1) (2025-08-12)
+
+
+### Bug Fixes
+
+* **placements:** fix side rail main element ([#983](https://github.com/Automattic/newspack-ads/issues/983)) ([5094654](https://github.com/Automattic/newspack-ads/commit/5094654a6625dc089661e092ec74ec782128d5f7))
+
+# [3.6.0](https://github.com/Automattic/newspack-ads/compare/v3.5.5...v3.6.0) (2025-08-11)
+
+
+### Features
+
+* **placements:** new sticky and side rail ([#975](https://github.com/Automattic/newspack-ads/issues/975)) ([18d6c78](https://github.com/Automattic/newspack-ads/commit/18d6c7881eb26d6489a2a9de5c5f7e27437f3ebf))
+
 ## [3.5.5](https://github.com/Automattic/newspack-ads/compare/v3.5.4...v3.5.5) (2025-07-28)